requirements.txt
clipspy
pysimplegui

Ewentualnie:
    Ubuntu:
        sudo apt-get install python3-tk
    Fedora:
        sudo dnf install python3-tkinter